+++ app/src/main/java/de/thomaskuenneth/benice/BeNiceCheckbox.kt
@@ -1,0 +1,30 @@
+package de.thomaskuenneth.benice
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.selection.toggleable
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.semantics.Role
+
+@Composable
+fun BeNiceCheckbox(
+    checked: Boolean,
+    text: String,
+    onCheckedChange: (Boolean) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier.toggleable(
+            value = checked, onValueChange = { onCheckedChange(!checked) }, role = Role.Checkbox
+        ), verticalAlignment = Alignment.CenterVertically
+    ) {
+        Checkbox(checked = checked, onCheckedChange = onCheckedChange)
+        Text(
+            text = text, style = MaterialTheme.typography.bodyLarge
+        )
+    }
+}