+++ whatsnew.md
@@ -1,6 +1,4 @@
 # What's new
 
-- Library and tools upgrades
 - UI refinements
-- Long pressing on an app icon has always opened a menu; this menu has a new entry called Copy names. It copies the package name of the app and the fully qualified class name of the activity to the system clipboard
-- 
+- Control number of app icon columns on small and medium screens